add authorization code and token auth
This is the working version of authentication. It allows a user to either run through a manual setup to recieve an authorization code, or use an existing refresh token to authenticate. The intention is for this to be as lightweight and modular as possible. Storing/handling tokens securely will be the responsibility of the user, not this package.